third_party/blink/web_tests/fast/dom/nodelist-item-parameter-expected.txt

Tests that the NodeList.item() argument is correctly validated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nodeList.__proto__ is NodeList.prototype
PASS nodeList.item(0).id is "a"
PASS nodeList.item(1).id is "b"
PASS nodeList.item(2) is null
PASS nodeList.item(-1) is null
PASS nodeList.item(-4294967295).id is "b"
PASS nodeList.item() threw exception TypeError: Failed to execute 'item' on 'NodeList': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
